Update README
Text Structuring: The text was organized into distinct sections with clear titles, which facilitates reading and understanding of the content.

Markdown Formatting: The content has been formatted in Markdown, which allows for a more pleasant and readable presentation when viewed on platforms such as GitHub.

Corrections and Additions: Some sections have been corrected and others have been added to provide more useful information to readers. For example, added information about debugging in Rails and accessing the database console.

English Translation: The final version includes an English translation, which makes the content accessible to a wider audience.

In short, the final version is more organized, presentable and informative than the initial version, providing a better experience for the reader.